+++ src/jobservice/job/impl/context.go
@@ -160,7 +160,7 @@ func getDBFromConfig(cfg map[string]interface{}) *models.Database {
 	database.Type = cfg[common.DatabaseType].(string)
 	postgresql := &models.PostGreSQL{}
 	postgresql.Host = cfg[common.PostGreSQLHOST].(string)
-	postgresql.Port = int(cfg[common.PostGreSQLPort].(int))
+	postgresql.Port = int(cfg[common.PostGreSQLPort].(float64))
 	postgresql.Username = cfg[common.PostGreSQLUsername].(string)
 	postgresql.Password = cfg[common.PostGreSQLPassword].(string)
 	postgresql.Database = cfg[common.PostGreSQLDatabase].(string)
